Refactor tooltip generation by replacing instance methods with direct imports from tools module
